Create src folder
